requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
Flask-WTF==1.2.1
WTForms==3.0.1
moviepy==1.0.3
Pillow>=10.0.0
python-dotenv==0.19.0